requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0